Fix: Drop all rows when logging out
