+++ app/hooks/useLogout.ts
@@ -3,6 +3,7 @@ import TrackPlayer from 'react-native-track-player';
 import {useStore} from '@boum/hooks';
 import {clearAllEncryptedValues} from '@boum/lib/encryptedStorage/encryptedStorage';
 import {Session} from '@boum/types';
+import {deleteAllRows, getDBConnection} from '@boum/lib/db';
 
 const initialSession: Session = {
   hostname: '',
@@ -25,7 +26,9 @@ const useLogout = async () => {
   try {
     await clearAllEncryptedValues();
     useStore.setState({session: initialSession});
+    const db = await getDBConnection();
     await TrackPlayer.reset();
+    deleteAllRows(db);
     return 'Succesfully logged out.';
   } catch (error) {
     return 'Error in saving User session.';
