+++ app/lib/db/index.ts
@@ -1,2 +1,2 @@
-export * from '@boum/lib/db/models';
+export * from '@boum/lib/db/customLists';
 export * from '@boum/lib/db/service';
